refactor: 헤더 컴포넌트 폴더 구조 이동
